test: mapped ltv validation tests
