+++ test-suites/_pool_core_erc20_borrow.spec.ts
@@ -259,5 +259,65 @@ describe("pToken/debtToken Borrow Event Accounting", () => {
         (await pool.getUserAccountData(user1.address)).erc721HealthFactor
       );
     });
+
+    it("TC-erc20-borrow-11 User 1 deposit DAI, DAI ltv drops to 0, then tries borrow (revert expected)", async () => {
+      testEnv = await loadFixture(testEnvFixture);
+      const {
+        pool,
+        dai,
+        weth,
+        users: [user1, user2],
+        configurator,
+        protocolDataProvider,
+      } = testEnv;
+
+      const daiAmount = await convertToCurrencyDecimals(dai.address, "10");
+      const wethAmount = await convertToCurrencyDecimals(weth.address, "10");
+      const borrowWethAmount = await convertToCurrencyDecimals(
+        weth.address,
+        "5"
+      );
+
+      await dai.connect(user1.signer).approve(pool.address, MAX_UINT_AMOUNT);
+      await weth.connect(user2.signer).approve(pool.address, MAX_UINT_AMOUNT);
+
+      await dai.connect(user1.signer)["mint(uint256)"](daiAmount);
+      await weth.connect(user2.signer)["mint(uint256)"](wethAmount);
+
+      await pool
+        .connect(user1.signer)
+        .supply(dai.address, daiAmount, user1.address, 0);
+      await pool
+        .connect(user2.signer)
+        .supply(weth.address, wethAmount, user2.address, 0);
+
+      // Set DAI LTV = 0
+      expect(
+        await configurator.configureReserveAsCollateral(
+          dai.address,
+          0,
+          8000,
+          10500
+        )
+      )
+        .to.emit(configurator, "CollateralConfigurationChanged")
+        .withArgs(dai.address, 0, 8000, 10500);
+      const ltv = (
+        await protocolDataProvider.getReserveConfigurationData(dai.address)
+      ).ltv;
+      expect(ltv).to.be.equal(0);
+
+      // Borrow all the weth because of issue in collateral needed.
+      await expect(
+        pool
+          .connect(user1.signer)
+          .borrow(weth.address, borrowWethAmount, 0, user1.address)
+      ).to.be.revertedWith(ProtocolErrors.LTV_VALIDATION_FAILED);
+
+      const userData = await pool.getUserAccountData(user1.address);
+      // failing here
+      // expect(userData.totalCollateralBase).to.be.eq(parseUnits("10", 8));
+      expect(userData.totalDebtBase).to.be.eq(0);
+    });
   });
 });
