+++ test-suites/_pool_core_erc20_withdraw.spec.ts
@@ -205,4 +205,119 @@ describe("pToken Withdraw Event Accounting", () => {
       );
     });
   });
+
+  context("LTV validation", () => {
+    let testEnv: TestEnv;
+    const {LTV_VALIDATION_FAILED} = ProtocolErrors;
+
+    before(async () => {
+      testEnv = await loadFixture(testEnvFixture);
+      const {
+        pool,
+        dai,
+        usdc,
+        weth,
+        users: [user1, user2],
+        configurator,
+        protocolDataProvider,
+      } = testEnv;
+
+      // User 1 deposits 10 Dai, 10 USDC, user 2 deposits 0.071 WETH
+      const daiAmount = await convertToCurrencyDecimals(dai.address, "10");
+      const usdcAmount = await convertToCurrencyDecimals(usdc.address, "10");
+      const wethAmount = await convertToCurrencyDecimals(weth.address, "0.071");
+
+      await dai.connect(user1.signer).approve(pool.address, MAX_UINT_AMOUNT);
+      await usdc.connect(user1.signer).approve(pool.address, MAX_UINT_AMOUNT);
+      await weth.connect(user2.signer).approve(pool.address, MAX_UINT_AMOUNT);
+
+      await dai.connect(user1.signer)["mint(uint256)"](daiAmount);
+      await usdc.connect(user1.signer)["mint(uint256)"](usdcAmount);
+      await weth.connect(user2.signer)["mint(uint256)"](wethAmount);
+
+      await pool
+        .connect(user1.signer)
+        .supply(dai.address, daiAmount, user1.address, 0);
+
+      await pool
+        .connect(user1.signer)
+        .supply(usdc.address, usdcAmount, user1.address, 0);
+
+      await pool
+        .connect(user2.signer)
+        .supply(weth.address, wethAmount, user2.address, 0);
+
+      // Set DAI LTV to 0
+      expect(
+        await configurator.configureReserveAsCollateral(
+          dai.address,
+          0,
+          8000,
+          10500
+        )
+      )
+        .to.emit(configurator, "CollateralConfigurationChanged")
+        .withArgs(dai.address, 0, 8000, 10500);
+
+      const ltv = (
+        await protocolDataProvider.getReserveConfigurationData(dai.address)
+      ).ltv;
+
+      expect(ltv).to.be.equal(0);
+
+      // borrow 0.000414 WETH
+      const borrowedAmount = await convertToCurrencyDecimals(
+        weth.address,
+        "0.000414"
+      );
+
+      expect(
+        await pool
+          .connect(user1.signer)
+          .borrow(weth.address, borrowedAmount, 0, user1.address)
+      );
+    });
+
+    it("TC-erc20-withdraw-07 User cannot withdraw an amount that leaves the borrow position uncovered based on LTV (revert expected)", async () => {
+      const {
+        pool,
+        usdc,
+        users: [user1],
+      } = testEnv;
+
+      const withdrawnAmount = await convertToCurrencyDecimals(
+        usdc.address,
+        "1"
+      );
+
+      await expect(
+        pool
+          .connect(user1.signer)
+          .withdraw(usdc.address, withdrawnAmount, user1.address)
+      ).to.be.revertedWith(LTV_VALIDATION_FAILED);
+    });
+
+    it("TC-erc20-withdraw-08 User can withdraw an amount that won't leave the borrow position uncovered based on LTV", async () => {
+      const {
+        pool,
+        dai,
+        pDai,
+        users: [user1],
+      } = testEnv;
+
+      const pDaiBalanceBefore = await pDai.balanceOf(user1.address);
+
+      const withdrawnAmount = await convertToCurrencyDecimals(dai.address, "1");
+
+      expect(
+        await pool
+          .connect(user1.signer)
+          .withdraw(dai.address, withdrawnAmount, user1.address)
+      );
+
+      const pDaiBalanceAfter = await pDai.balanceOf(user1.address);
+
+      expect(pDaiBalanceAfter).to.be.eq(pDaiBalanceBefore.sub(withdrawnAmount));
+    });
+  });
 });
